+++ spec/integration/zanox_integration_spec.rb
@@ -17,10 +17,11 @@ RSpec.describe "Zanox Integration" do
     expect(page).to have_content "var zx_category = 'cake decorating';\nvar zx_amount = '5.90';"
   end
 
-  it 'should include the track event' do
-    expect(page).to have_xpath "//script[contains(@src,'12345H123456789&mode=[[1]]&CustomerID=[[123456]]&OrderID=[[DEFC-4321]]&CurrencySymbol=[[EUR]]&TotalPrice=[[150.00]]')]"
+  it 'should include the sale event' do
+    expect(page).to have_xpath "//script[contains(@src,'pps/?12345H123456789&mode=[[1]]&CustomerID=[[123456]]&OrderID=[[DEFC-4321]]&CurrencySymbol=[[EUR]]&TotalPrice=[[150.00]]')]"
+  end
+
+  it 'should include the lead event' do
+    expect(page).to have_xpath "//script[contains(@src,'ppl/?12345H123456789&mode=[[1]]&CustomerID=[[654321]]')]"
   end
 end
-
-
-